Check if the post is an order before deleting it

+++ src/Api/Order.php
@@ -136,7 +136,10 @@ class Order extends AbstractEndpoint
 
 			if ( is_wp_error( $order ) ) {
 				// Because we can have errors in $order, delete the order and return the error.
-				wp_delete_post( $order_id );
+				if ( in_array( get_post( $order_id )->post_type, wc_get_order_types() ) ) {
+					wp_delete_post( $order_id );
+				}
+
 				return $order;
 			}
 		}
